chore: clean up debug files and screenshots
- Removed temporary authentication debug scripts
- Removed test auth state files
- Added screenshot from successful test run
- Cleanup after successful E2E authentication fixes

E2E test status: 16/25 tests passing with single-user strategy
